jackson parse test - get subcategory
